+++ mods/CreateDragonPlus/src/main/java/plus/dragons/createdragonsplus/common/fluids/dye/DyeVariantRegistry.java
@@ -19,7 +19,6 @@
 package plus.dragons.createdragonsplus.common.fluids.dye;
 
 import java.util.Collection;
-import java.util.Comparator;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;